+++ lib/src/settings.rs
@@ -243,19 +243,23 @@ impl UserSettings {
     }
 
     pub fn commit_node_template(&self) -> String {
-        self.node_template_for_key(
-            "templates.log_node",
-            r#"if(self, if(current_working_copy, "@", "◉"), "◌")"#,
-            r#"if(self, if(current_working_copy, "@", "o"), ".")"#,
-        )
+        let key = match self.graph_style().as_str() {
+            "ascii" | "ascii-large" => "templates.log_node_ascii",
+            _ => "templates.log_node",
+        };
+        self.config
+            .get_string(key)
+            .expect("Node templates should exist in bundled config!")
     }
 
     pub fn op_node_template(&self) -> String {
-        self.node_template_for_key(
-            "templates.op_log_node",
-            r#"if(current_operation, "@", "◉")"#,
-            r#"if(current_operation, "@", "o")"#,
-        )
+        let key = match self.graph_style().as_str() {
+            "ascii" | "ascii-large" => "templates.op_log_node_ascii",
+            _ => "templates.op_log_node",
+        };
+        self.config
+            .get_string(key)
+            .expect("Node templates should exist in bundled config!")
     }
 
     pub fn max_new_file_size(&self) -> Result<u64, config::ConfigError> {
@@ -279,14 +283,6 @@ impl UserSettings {
 
     pub fn sign_settings(&self) -> SignSettings {
         SignSettings::from_settings(self)
-    }
-
-    fn node_template_for_key(&self, key: &str, fallback: &str, ascii_fallback: &str) -> String {
-        let symbol = self.config.get_string(key);
-        match self.graph_style().as_str() {
-            "ascii" | "ascii-large" => symbol.unwrap_or_else(|_| ascii_fallback.to_owned()),
-            _ => symbol.unwrap_or_else(|_| fallback.to_owned()),
-        }
     }
 }